feat: integrate Piscina for worker-based evaluation processing
- Added Piscina for handling evaluation tasks in a worker thread to improve performance.
- Refactored EvaluationService to dispatch evaluation tasks to the worker.
- Created a new evaluation.worker.js file to handle the evaluation logic.
- Updated package.json to include Piscina as a dependency.
- Removed synchronous evaluation logic from EvaluationService and moved it to the worker.
- Enhanced logging for better traceability during evaluation processing.
