+++ package.json
@@ -47,6 +47,7 @@
     "mongoose": "^8.8.3",
     "npm": "^10.9.0",
     "openai": "^4.73.1",
+    "piscina": "^5.0.0",
     "playwright": "^1.49.1",
     "playwright-extra": "^4.3.6",
     "playwright-extra-plugin-stealth": "^0.0.1",
@@ -66,7 +67,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "vitest run",
-    "eject": "react-scripts eject",    "start-server": "node server/server.js",
+    "eject": "react-scripts eject",
+    "start-server": "node server/server.js",
     "dev": "concurrently \"npm run start-server\" \"npm start\"",
     "lint": "cross-env NODE_ENV=development dotenv -e .env eslint \"src/**/*.{js,jsx}\""
   },
